#4 - configured FluentValidation lib and implemented for AuthController Dtos
