+++ WhiskyMan.Backend/WhiskyMan.Entities/BottleDescription.cs
@@ -12,15 +12,15 @@ namespace WhiskyMan.Entities
         public int Id { get; set; }
 
         [Required]
-        [MaxLength(100)]
+        [MaxLength(EntitiesConfig.NameLength)]
         public string Name { get; set; }
         [Required]
-        [MaxLength(100)]
+        [MaxLength(EntitiesConfig.NameLength)]
         public string Distillery { get; set; }
         public int Age { get; set; }
         public decimal Voltage { get; set; }
         [Required]
-        [MaxLength(300)]
+        [MaxLength(EntitiesConfig.UrlLength)]
         public string PictureUrl { get; set; }
 
         public Collection<Bottle> Bottles { get; set; }
