+++ WhiskyMan.Backend/WhiskyMan.Entities/EntitiesConfig.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace WhiskyMan.Entities
+{
+    public static class EntitiesConfig
+    {
+        public const int UrlLength = 500;
+        public const int NameLength = 50;
+        public const int EmailLength = 50;
+        public const int PasswdLength = 100;
+    }
+}
