--- WhiskyMan.Backend/WhiskyMan.Models/Dtos/User/UserForLogin.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace WhiskyMan.Models.Dtos.User
-{
-    public class UserForLogin
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
