--- WhiskyMan.Backend/WhiskyMan.Models/Dtos/User/UserForRegister.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace WhiskyMan.Models.Dtos.User
-{
-    public class UserForRegister
-    {
-        public string Username { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public string PictureUrl { get; set; }
-    }
-}
