+++ WhiskyMan.Backend/WhiskyMan.Models/WhiskyMan.Models.csproj
@@ -4,4 +4,12 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="FluentValidation" Version="9.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\WhiskyMan.Entities\WhiskyMan.Entities.csproj" />
+  </ItemGroup>
+
 </Project>
